requirements.txt
asgiref==3.3.1
bson==0.5.8
certifi==2020.12.5
chardet==4.0.0
dataclasses==0.6
Django==3.0.5
djangorestframework==3.12.2
djongo==1.3.4
dnspython==2.0.0
idna==2.10
pymongo==3.11.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.2.4
urllib3==1.26.3
